pyproject.toml
[tool.poetry]
name = "vis-action-seg"
version = "0.1.0"
description = "Library for visualizing action segmentation results."
authors = ["yiskw713"]
license = "MIT License"
readme = "README.md"
packages = [{include = "vis_action_seg"}]

[tool.poetry.dependencies]
python = "^3.9"
pillow = "^9.2.0"
numpy = "^1.23.4"


[tool.poetry.group.dev.dependencies]
mypy = "^0.982"
isort = "^5.10.1"
flake8 = "^5.0.4"
black = "^22.10.0"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
